LUCENE-8621: Refactor LatLonShape, XYShape, and all query and utility classes from sandbox to core
