+++ lucene/core/src/java/org/apache/lucene/document/LatLonShape.java
@@ -52,8 +52,6 @@ import static org.apache.lucene.geo.GeoEncodingUtils.encodeLongitude;
  * and 8.381903171539307E-8 for longitude).
  * @see PointValues
  * @see LatLonDocValuesField
- *
- * @lucene.experimental
  */
 public class LatLonShape {
