+++ lucene/core/src/java/org/apache/lucene/document/LatLonShapePointQuery.java
@@ -19,26 +19,22 @@ package org.apache.lucene.document;
 import java.util.Arrays;
 
 import org.apache.lucene.geo.Component2D;
+import org.apache.lucene.geo.GeoEncodingUtils;
 import org.apache.lucene.geo.Point2D;
-import org.apache.lucene.geo.XYEncodingUtils;
 import org.apache.lucene.index.PointValues.Relation;
 import org.apache.lucene.util.NumericUtils;
-
-import static org.apache.lucene.geo.XYEncodingUtils.decode;
 
 /**
  * Finds all previously indexed shapes that intersect the specified bounding box.
  *
  * <p>The field must be indexed using
- * {@link XYShape#createIndexableFields} added per document.
- *
- *  @lucene.experimental
+ * {@link LatLonShape#createIndexableFields} added per document.
  **/
-final class XYShapePointQuery extends ShapeQuery {
+final class LatLonShapePointQuery extends ShapeQuery {
   final Component2D point2D;
-  final float[][] points;
+  final double[][] points;
 
-  public XYShapePointQuery(String field, ShapeField.QueryRelation queryRelation, float[][] points) {
+  public LatLonShapePointQuery(String field, ShapeField.QueryRelation queryRelation, double[][] points) {
     super(field, queryRelation);
     this.points = points;
     this.point2D = Point2D.create(points);
@@ -47,13 +43,13 @@ final class XYShapePointQuery extends ShapeQuery {
   @Override
   protected Relation relateRangeBBoxToQuery(int minXOffset, int minYOffset, byte[] minTriangle,
                                             int maxXOffset, int maxYOffset, byte[] maxTriangle) {
-    double minY = XYEncodingUtils.decode(NumericUtils.sortableBytesToInt(minTriangle, minYOffset));
-    double minX = XYEncodingUtils.decode(NumericUtils.sortableBytesToInt(minTriangle, minXOffset));
-    double maxY = XYEncodingUtils.decode(NumericUtils.sortableBytesToInt(maxTriangle, maxYOffset));
-    double maxX = XYEncodingUtils.decode(NumericUtils.sortableBytesToInt(maxTriangle, maxXOffset));
+    double minLat = GeoEncodingUtils.decodeLatitude(NumericUtils.sortableBytesToInt(minTriangle, minYOffset));
+    double minLon = GeoEncodingUtils.decodeLongitude(NumericUtils.sortableBytesToInt(minTriangle, minXOffset));
+    double maxLat = GeoEncodingUtils.decodeLatitude(NumericUtils.sortableBytesToInt(maxTriangle, maxYOffset));
+    double maxLon = GeoEncodingUtils.decodeLongitude(NumericUtils.sortableBytesToInt(maxTriangle, maxXOffset));
 
     // check internal node against query
-    return point2D.relate(minX, maxX, minY, maxY);
+    return point2D.relate(minLon, maxLon, minLat, maxLat);
   }
 
   /** returns true if the query matches the encoded triangle */
@@ -61,20 +57,20 @@ final class XYShapePointQuery extends ShapeQuery {
   protected boolean queryMatches(byte[] t, ShapeField.DecodedTriangle scratchTriangle, ShapeField.QueryRelation queryRelation) {
     ShapeField.decodeTriangle(t, scratchTriangle);
 
-    double aY = decode(scratchTriangle.aY);
-    double aX = decode(scratchTriangle.aX);
-    double bY = decode(scratchTriangle.bY);
-    double bX = decode(scratchTriangle.bX);
-    double cY = decode(scratchTriangle.cY);
-    double cX = decode(scratchTriangle.cX);
+    double alat = GeoEncodingUtils.decodeLatitude(scratchTriangle.aY);
+    double alon = GeoEncodingUtils.decodeLongitude(scratchTriangle.aX);
+    double blat = GeoEncodingUtils.decodeLatitude(scratchTriangle.bY);
+    double blon = GeoEncodingUtils.decodeLongitude(scratchTriangle.bX);
+    double clat = GeoEncodingUtils.decodeLatitude(scratchTriangle.cY);
+    double clon = GeoEncodingUtils.decodeLongitude(scratchTriangle.cX);
 
     switch (queryRelation) {
       case INTERSECTS:
-        return point2D.relateTriangle(aX, aY, bX, bY, cX, cY) != Relation.CELL_OUTSIDE_QUERY;
+        return point2D.relateTriangle(alon, alat, blon, blat, clon, clat) != Relation.CELL_OUTSIDE_QUERY;
       case WITHIN:
-        return point2D.relateTriangle(aX, aY, bX, bY, cX, cY) == Relation.CELL_INSIDE_QUERY;
+        return point2D.relateTriangle(alon, alat, blon, blat, clon, clat) == Relation.CELL_INSIDE_QUERY;
       case DISJOINT:
-        return point2D.relateTriangle(aX, aY, bX, bY, cX, cY) == Relation.CELL_OUTSIDE_QUERY;
+        return point2D.relateTriangle(alon, alat, blon, blat, clon, clat) == Relation.CELL_OUTSIDE_QUERY;
       default:
         throw new IllegalArgumentException("Unsupported query type :[" + queryRelation + "]");
     }
@@ -84,14 +80,14 @@ final class XYShapePointQuery extends ShapeQuery {
   protected Component2D.WithinRelation queryWithin(byte[] t, ShapeField.DecodedTriangle scratchTriangle) {
     ShapeField.decodeTriangle(t, scratchTriangle);
 
-    double aY = decode(scratchTriangle.aY);
-    double aX = decode(scratchTriangle.aX);
-    double bY = decode(scratchTriangle.bY);
-    double bX = decode(scratchTriangle.bX);
-    double cY = decode(scratchTriangle.cY);
-    double cX = decode(scratchTriangle.cX);
+    double alat = GeoEncodingUtils.decodeLatitude(scratchTriangle.aY);
+    double alon = GeoEncodingUtils.decodeLongitude(scratchTriangle.aX);
+    double blat = GeoEncodingUtils.decodeLatitude(scratchTriangle.bY);
+    double blon = GeoEncodingUtils.decodeLongitude(scratchTriangle.bX);
+    double clat = GeoEncodingUtils.decodeLatitude(scratchTriangle.cY);
+    double clon = GeoEncodingUtils.decodeLongitude(scratchTriangle.cX);
 
-    return point2D.withinTriangle(aX, aY, scratchTriangle.ab, bX, bY, scratchTriangle.bc, cX, cY, scratchTriangle.ca);
+    return point2D.withinTriangle(alon, alat, scratchTriangle.ab, blon, blat, scratchTriangle.bc, clon, clat, scratchTriangle.ca);
   }
 
   @Override
@@ -101,7 +97,7 @@ final class XYShapePointQuery extends ShapeQuery {
 
   @Override
   protected boolean equalsTo(Object o) {
-    return super.equalsTo(o) && Arrays.equals(points, ((XYShapePointQuery)o).points);
+    return super.equalsTo(o) && Arrays.equals(points, ((LatLonShapePointQuery)o).points);
   }
 
   @Override
