+++ lucene/core/src/java/org/apache/lucene/document/ShapeQuery.java
@@ -56,8 +56,6 @@ import org.apache.lucene.util.FixedBitSet;
  *   <li>{@link XYShape#newPolygonQuery newPolygonQuery()} for matching cartesian shapes that have some {@link QueryRelation} with a polygon.
  * </ul>
  * <p>
- *
- *  @lucene.experimental
  **/
 abstract class ShapeQuery extends Query {
   /** field name */
@@ -203,6 +201,7 @@ abstract class ShapeQuery extends Query {
     return sameClassAs(o) && equalsTo(o);
   }
 
+  /** class specific equals check */
   protected boolean equalsTo(Object o) {
     return Objects.equals(field, ((ShapeQuery)o).field) && this.queryRelation == ((ShapeQuery)o).queryRelation;
   }
