+++ lucene/core/src/java/org/apache/lucene/document/XYShape.java
@@ -48,8 +48,6 @@ import static org.apache.lucene.geo.XYEncodingUtils.encode;
  * original {@code double} values.
  * @see PointValues
  * @see LatLonDocValuesField
- *
- * @lucene.experimental
  */
 public class XYShape {
