+++ lucene/core/src/java/org/apache/lucene/geo/Line2D.java
