+++ lucene/core/src/java/org/apache/lucene/geo/Point2D.java
