+++ lucene/core/src/java/org/apache/lucene/geo/Rectangle2D.java
