+++ lucene/core/src/java/org/apache/lucene/geo/SimpleWKTShapeParser.java
