+++ lucene/core/src/java/org/apache/lucene/geo/Tessellator.java
@@ -66,7 +66,7 @@ import static org.apache.lucene.geo.GeoUtils.orient;
  * TORTIOUS ACTION, ARISING OUT OF OR IN CONNECTION WITH THE USE OR PERFORMANCE OF
  * THIS SOFTWARE.
  *
- * @lucene.experimental
+ * @lucene.internal
  */
 final public class Tessellator {
   // this is a dumb heuristic to control whether we cut over to sorted morton values
