+++ lucene/core/src/java/org/apache/lucene/geo/XYPolygon2D.java
