+++ lucene/core/src/java/org/apache/lucene/geo/XYRectangle.java
