+++ lucene/core/src/java/org/apache/lucene/geo/XYRectangle2D.java
