+++ lucene/core/src/test/org/apache/lucene/document/BaseLatLonShapeTestCase.java
