+++ lucene/core/src/test/org/apache/lucene/document/BaseShapeEncodingTestCase.java
