+++ lucene/core/src/test/org/apache/lucene/document/BaseShapeTestCase.java
