+++ lucene/core/src/test/org/apache/lucene/document/BaseXYShapeTestCase.java
