+++ lucene/core/src/test/org/apache/lucene/document/TestLatLonLineShapeQueries.java
