+++ lucene/core/src/test/org/apache/lucene/document/TestLatLonMultiLineShapeQueries.java
