+++ lucene/core/src/test/org/apache/lucene/document/TestLatLonMultiPointShapeQueries.java
