+++ lucene/core/src/test/org/apache/lucene/document/TestLatLonMultiPolygonShapeQueries.java
