+++ lucene/core/src/test/org/apache/lucene/document/TestLatLonPointShapeQueries.java
