+++ lucene/core/src/test/org/apache/lucene/document/TestLatLonPolygonShapeQueries.java
