+++ lucene/core/src/test/org/apache/lucene/document/TestLatLonShape.java
