+++ lucene/core/src/test/org/apache/lucene/document/TestLatLonShapeEncoding.java
