+++ lucene/core/src/test/org/apache/lucene/document/TestXYLineShapeQueries.java
