+++ lucene/core/src/test/org/apache/lucene/document/TestXYMultiLineShapeQueries.java
