+++ lucene/core/src/test/org/apache/lucene/document/TestXYMultiPointShapeQueries.java
