+++ lucene/core/src/test/org/apache/lucene/document/TestXYMultiPolygonShapeQueries.java
