+++ lucene/core/src/test/org/apache/lucene/document/TestXYPointShapeQueries.java
