+++ lucene/core/src/test/org/apache/lucene/document/TestXYPolygonShapeQueries.java
