+++ lucene/core/src/test/org/apache/lucene/document/TestXYShape.java
