+++ lucene/core/src/test/org/apache/lucene/document/TestXYShapeEncoding.java
