+++ lucene/core/src/test/org/apache/lucene/geo/TestPoint2D.java
