+++ lucene/core/src/test/org/apache/lucene/geo/TestRectangle2D.java
