+++ lucene/core/src/test/org/apache/lucene/geo/TestSimpleWKTShapeParsing.java
