+++ lucene/core/src/test/org/apache/lucene/geo/TestXYRectangle2D.java
